// RestaurantManagementSystem/frontend/src/interfaces/OrderMealsTable.ts
export interface OrderMealTableInterface {
    mealName: string;
    mealPrice: number;
    quantity: number;
    totalPrice: string;
    time: string;
}

export interface OrderMealTableResponse {
    content: OrderMealTableInterface[];
    pageable: {
        pageNumber: number;
        pageSize: number;
        sort: {
            empty: boolean;
            sorted: boolean;
            unsorted: boolean;
        };
        offset: number;
        unpaged: boolean;
        paged: boolean;
    };
    last: boolean;
    totalElements: number;
    totalPages: number;
    size: number;
    number: number;
    sort: {
        empty: boolean;
        sorted: boolean;
        unsorted: boolean;
    };
    first: boolean;
    numberOfElements: number;
    empty: boolean;
}